Make Multi part pie chart work.
