+++ Sources/SwiftUICharts/PieChart/Models/MultiLayer/MultiLayerPieChartData.swift
@@ -60,87 +60,3 @@ public final class MultiLayerPieChartData: MultiPieChartDataProtocol {
     public typealias DataPoint = MultiPieDataPoint
     public typealias CTStyle   = PieChartStyle
 }
-
-
-
-// MARK: - Data Sets
-public struct MultiPieDataSet: SingleDataSet {
-    
-    public var id: UUID = UUID()
-    public var dataPoints  : [MultiPieDataPoint]
-    
-    public init(dataPoints: [MultiPieDataPoint]) {
-        self.dataPoints = dataPoints
-    }
-    
-    public typealias DataPoint = MultiPieDataPoint
-
-}
-
-
-
-// MARK: - Data Point
-public struct MultiPieDataPoint: CTPieDataPoint {
-    
-    public var id: UUID = UUID()
-    // CTPieDataPoint
-    public var startAngle  : Double = 0
-    public var amount      : Double = 0
-    // CTChartDataPoint
-    public var value            : Double
-    public var pointDescription : String?
-    public var date             : Date?
-    
-    public var colour           : Color
-    
-    public var layerDataPoints  : [MultiPieDataPoint]?
-    
-    public init(value           : Double,
-                pointDescription: String?   = nil,
-                date            : Date?     = nil,
-                colour          : Color     = Color.red,
-                layerDataPoints : [MultiPieDataPoint]? = nil
-    ) {
-        self.value              = value
-        self.pointDescription   = pointDescription
-        self.date               = date
-        self.colour             = colour
-        
-        self.layerDataPoints    = layerDataPoints
-        
-    }
-
-}
-
-// MARK: - View
-
-public struct MultiLayerPie<ChartData>: View where ChartData: MultiLayerPieChartData {
-    
-    @ObservedObject var chartData: ChartData
-
-    public init(chartData: ChartData) {
-        self.chartData = chartData
-    }
-    
-    public var body: some View {
-        
-        ZStack {
-            ForEach(chartData.dataSets.dataPoints, id: \.self) { data in
-                PieSegmentShape(id:         data.id,
-                                startAngle: data.startAngle,
-                                amount:     data.amount)
-                    .fill(data.colour)
-                
-                if let points = data.layerDataPoints {
-                    ForEach(points, id: \.self) { point in
-                        DoughnutSegmentShape(id:         point.id,
-                                             startAngle: point.startAngle,
-                                             amount:     point.amount)
-                            .strokeBorder(point.colour, lineWidth: 60)
-                    }
-                    
-                }
-            }
-        }
-    }
-}
