+++ Sources/SwiftUICharts/PieChart/Shapes/PieSegmentShape.swift
@@ -13,6 +13,8 @@ internal struct PieSegmentShape: Shape, Identifiable {
     var startAngle  : Double
     var amount      : Double
     
+
+    
     internal func path(in rect: CGRect) -> Path {
         
         let radius = min(rect.width, rect.height) / 2
